Update TOP10.java
Problem

Chef has been working hard to compete in MasterChef.
He is ranked XX out of all contestants. However, only 1010 contestants would be selected for the finals.

Check whether Chef made it to the top 1010 or not?
Input Format

    The first line of input will contain a single integer TT, denoting the number of test cases.
    Each test case consists of one integers XX — the current rank of Chef.

Output Format

For each test case, output on a new line, YES, if Chef made it to the top 1010 and NO otherwise.

Each character of the output may be printed in either uppercase or lowercase. That is, the strings NO, no, nO, and No will be treated as equivalent.
Constraints

    1≤T≤1001≤T≤100
    1≤X≤1001≤X≤100

Sample 1:
Input
Output

4
15
10
1
50

NO
YES
YES
NO

Explanation:

Test case 11: Chef's rank is 1515 which is greater than 1010. Thus, Chef did not make it to the top 1010.

Test case 22: Chef's rank is 1010 which is equal to 1010. Thus, Chef made it to the top 1010.

Test case 33: Chef made it to the top 1010, as his rank is 11.

Test case 44: Chef did not make it to the top 1010 as his rank is 5050.